+++ package.json
@@ -16,7 +16,7 @@
     "@next/mdx": "^15.0.3",
     "@payloadcms/db-postgres": "^3.1.0",
     "@payloadcms/next": "^3.1.0",
-    "@payloadcms/richtext-lexical": "^3.1.0",
+    "@payloadcms/richtext-lexical": "3.1.0",
     "@radix-ui/react-dialog": "^1.1.2",
     "@radix-ui/react-dropdown-menu": "^2.1.2",
     "@radix-ui/react-label": "^2.1.0",
