+++ src/app/api/questions/route.ts
@@ -1,28 +1,37 @@
-// src/app/api/questions/route.ts
-import { NextResponse } from 'next/server';
-import { getPayload } from 'payload';
-import config from '@/payload.config';
+import { NextResponse } from "next/server";
+import { getPayload } from "payload";
+import config from "@/payload.config";// Import the Questions type
 
 export async function GET() {
   try {
     const payload = await getPayload({ config });
 
     const { docs: questions } = await payload.find({
-      collection: 'questions',
-      depth: 2, 
-      page: 1,
-      limit: 10,
-      pagination: false, 
-      where: {}, 
-      sort: '-title',
-      fallbackLocale: false,
-      overrideAccess: false,
-      showHiddenFields: true, 
-    });
+      collection: "questions",
+      depth: 3,
+      populate: {
+        topics: {
+          name: true,
+        },
+        companies: {
+          name: true,
+        },
+        questions: {
+          title: true,
+          author: true,
+          body: true,
+          difficulty: true,
 
-    return NextResponse.json(questions); 
-  } catch (error: any) {
-    console.error('Error fetching questions:', error);
-    return NextResponse.json({ error: error.message || 'An error occurred while fetching questions.' }, { status: 500 });
+        }
+      },
+    });
+    console.log(JSON.stringify(questions));
+    return NextResponse.json(questions);
+  } catch (error: unknown) {
+    console.error("Error fetching questions:", error);
+    return NextResponse.json(
+      { error: (error as Error)?.message || "An error occurred while fetching questions." },
+      { status: 500 }
+    );
   }
 }
