+++ src/app/collections/Questions.ts
@@ -1,3 +1,4 @@
+import { HTMLConverterFeature, lexicalEditor, lexicalHTML } from '@payloadcms/richtext-lexical';
 import { CollectionConfig } from 'payload';
 
 
@@ -45,7 +46,16 @@ const Questions: CollectionConfig = {
     {
       name: 'body',
       type: 'richText',
+      editor: lexicalEditor({
+        features: ({ defaultFeatures }) => [
+          ...defaultFeatures,
+          // The HTMLConverter Feature is the feature which manages the HTML serializers.
+          // If you do not pass any arguments to it, it will use the default serializers.
+          HTMLConverterFeature({}),
+        ],
+      }),
     },
+    lexicalHTML('body', { name: 'body_html' }),
     {
       name: 'tests',
       type: 'array',
